added winning and loosing sound

+++ CandyJackpot/config.h
@@ -3,7 +3,7 @@
 
 // Pin Definitions
 const uint8_t buttonPin = 2;
-const int speakerPin = 3;
+const uint8_t speakerPin = 3;
 const int dispenserServoPin = 9;
 const int leverServoPin = 10;
